Fix return type annotation of _aggregate_rss_bytes to int
Addresses Copilot PR review comment: the function sums psutil's
integer .rss values and returns a byte count, not a float — the
GB conversion happens later in ResourceSampler._sample.

+++ src/ivory/utils/resource_sampler.py
@@ -3,7 +3,7 @@ import threading
 import psutil
 
 
-def _aggregate_rss_bytes(process: psutil.Process) -> float:
+def _aggregate_rss_bytes(process: psutil.Process) -> int:
     """Sum RSS of `process` and all its live children, recursively.
 
     Silently skips any process that disappears or is inaccessible mid-poll.
